// src/contracts/generated/openapi-registry.ts
export const OPENAPI_OPERATION_PATHS = {
  index_krx_dd_trd: "/svc/apis/idx/krx_dd_trd",
  index_kospi_dd_trd: "/svc/apis/idx/kospi_dd_trd",
  index_kosdaq_dd_trd: "/svc/apis/idx/kosdaq_dd_trd",
  index_bon_dd_trd: "/svc/apis/idx/bon_dd_trd",
  index_drvprod_dd_trd: "/svc/apis/idx/drvprod_dd_trd",
  stock_stk_bydd_trd: "/svc/apis/sto/stk_bydd_trd",
  stock_ksq_bydd_trd: "/svc/apis/sto/ksq_bydd_trd",
  stock_knx_bydd_trd: "/svc/apis/sto/knx_bydd_trd",
  stock_sw_bydd_trd: "/svc/apis/sto/sw_bydd_trd",
  stock_sr_bydd_trd: "/svc/apis/sto/sr_bydd_trd",
  stock_stk_isu_base_info: "/svc/apis/sto/stk_isu_base_info",
  stock_ksq_isu_base_info: "/svc/apis/sto/ksq_isu_base_info",
  stock_knx_isu_base_info: "/svc/apis/sto/knx_isu_base_info",
  etp_etf_bydd_trd: "/svc/apis/etp/etf_bydd_trd",
  etp_etn_bydd_trd: "/svc/apis/etp/etn_bydd_trd",
  etp_elw_bydd_trd: "/svc/apis/etp/elw_bydd_trd",
  bond_kts_bydd_trd: "/svc/apis/bon/kts_bydd_trd",
  bond_bnd_bydd_trd: "/svc/apis/bon/bnd_bydd_trd",
  bond_smb_bydd_trd: "/svc/apis/bon/smb_bydd_trd",
  derivative_fut_bydd_trd: "/svc/apis/drv/fut_bydd_trd",
  derivative_eqsfu_stk_bydd_trd: "/svc/apis/drv/eqsfu_stk_bydd_trd",
  derivative_eqkfu_ksq_bydd_trd: "/svc/apis/drv/eqkfu_ksq_bydd_trd",
  derivative_opt_bydd_trd: "/svc/apis/drv/opt_bydd_trd",
  derivative_eqsop_bydd_trd: "/svc/apis/drv/eqsop_bydd_trd",
  derivative_eqkop_bydd_trd: "/svc/apis/drv/eqkop_bydd_trd",
  commodity_oil_bydd_trd: "/svc/apis/gen/oil_bydd_trd",
  commodity_gold_bydd_trd: "/svc/apis/gen/gold_bydd_trd",
  commodity_ets_bydd_trd: "/svc/apis/gen/ets_bydd_trd",
  esg_sri_bond_info: "/svc/apis/esg/sri_bond_info",
  esg_esg_etp_info: "/svc/apis/esg/esg_etp_info",
  esg_esg_index_info: "/svc/apis/esg/esg_index_info",
} as const;